+++ src/snowflake/cli/_plugins/streamlit/commands.py
@@ -37,6 +37,7 @@ from snowflake.cli.api.commands.decorators import (
     with_project_definition,
 )
 from snowflake.cli.api.commands.flags import (
+    PruneOption,
     ReplaceOption,
     entity_argument,
     identifier_argument,
@@ -136,6 +137,7 @@ def streamlit_deploy(
         help="Replaces the Streamlit app if it already exists. It only uploads new and overwrites existing files, "
         "but does not remove any files already on the stage."
     ),
+    prune: bool = PruneOption(),
     entity_id: str = entity_argument("streamlit"),
     open_: bool = OpenOption,
     **options,
@@ -168,6 +170,7 @@ def streamlit_deploy(
         streamlit=streamlit,
         streamlit_project_paths=streamlit_project_paths,
         replace=replace,
+        prune=prune,
     )
 
     if open_: