+++ RELEASE-NOTES.md
@@ -19,6 +19,17 @@
 ## Deprecations
 
 ## New additions
+* Added `--prune` flag to `deploy` commands, which removes files that exist in the stage,
+  but not in the local filesystem.
+
+## Fixes and improvements
+
+
+# v3.6.0
+
+## Deprecations
+
+## New additions
 * Added new `snow auth keypair` commands:
     * `setup` - generate key pair, set public key for the user in Snowflake and update/create connection.
     * `rotate` - rotate keys for connection.
@@ -33,6 +44,8 @@
 * Fix for incompatible options in `snow spcs compute-pool` commands didn't raise error.
 * Change binary builds to embed whole Python environment.
 * Fixed recursive copying to stage for unbalanced directory trees.
+* Fixed checking for new CLI version.
+* Added file execution log in stage and git commands.
 
 # v3.5.0